# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Pod3 ASDA Project CRUD",
      "description": "Pod3 ASDA Project Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "3.88.168.68:8000",
    "basePath": "/",
    "tags": [
      {
        "conversations": "text conversations between user and chatbot",
        "components": "API for conversations between user and chatbot"
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],

    "paths": {
        "/initialize": {
            "post": {
                "summary": "Initalise the chain.",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "initInfo",
                        "description": "User and conversation info to initialise the chain.",
                        "schema": {
                            "$ref": "#/definitions/InitialiseInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref":"#/definitions/ConversationOutput"
                        }
                    }
                }    
            }
        },
        "/conversation": {
            "post": {
                "summary": "A message is posted and a response is received",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "promptInfo",
                        "description": "What the user sends to the chatbot.",
                        "schema": {
                            "$ref":"#/definitions/ConversationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref":"#/definitions/ConversationOutput"
                        }
                    }
                }                    
            }
        }
    },

    "definitions": {
        "InitialiseInput": {
            "required": ["userID", "conversationID"],
            "type": "object",
            "properties": {
                "userID": {
                "type": "integer"
                },
                "conversationID": {
                "type": "integer"
                }
            }
        },
        "ConversationInput": {
            "required": ["prompt", "userID", "conversationID", "firstMsg"],
            "type": "object",
            "properties": {
                "prompt": {
                "type": "string"
                },
                "userID": {
                "type": "integer"
                },
                "conversationID": {
                "type": "integer"
                }
            }
        },
        "ConversationOutput": {
            "required": ["output", "chatHistory"],
            "type": "object",
            "properties": {
                "output": {
                    "type": "object",
                    "properties": {
                        "text": {
                            "type": "string"
                        }
                    }
                },
                "chatHistory": {
                "type": "array",
                "items": {
                    "type": "object",
                    "properties": {
                    "role": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    }
                    }
                }
            }
        }
        }
    }
}